Rewriting of Problem Definition and Final conclusion of Motivation
Rewriting the Problem Definition section of the Introduction chapter to clarify what are the main problems and the focus that will be taken into account to solve the problem. Also, a final paragraph was added to the motivation section to explain that OutSystems wants to provide a faster, effective, and reliable solution to build queries.